pyproject.toml
[project]
name = "facilitatorbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.1",
    "python-dotenv>=1.1.0",
]
